Cadastro Veterinario

+++ ClinicaPet/src/main/java/com/Veterinaria/ClinicaPet/security/repositories/EnderecoRepository.java
@@ -1,11 +1,9 @@
 package com.veterinaria.clinicapet.security.repositories;
 
+import com.veterinaria.clinicapet.security.entities.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.veterinaria.clinicapet.security.entities.Endereco;
-
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
-    
+public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
 }
